bench: move storage-options env helpers to own module
