+++ benches/utils/mod.rs
@@ -20,6 +20,7 @@ pub mod markdown;
 pub mod report;
 pub mod rss;
 pub mod storage_meter;
+pub mod storage_options;
 pub mod tiers;
 
 pub mod superfile;
