+++ benches/utils/tiers.rs
@@ -11,7 +11,6 @@
 //! `INFINO_REAL_S3_BUCKET`, `azure` reads `INFINO_REAL_AZURE_CONTAINER`.
 
 use std::{
-    collections::HashMap,
     net::SocketAddr,
     sync::{Arc, OnceLock},
 };
@@ -30,6 +29,8 @@ use s3s::{auth::SimpleAuth, service::S3ServiceBuilder};
 use s3s_fs::FileSystem;
 use tempfile::TempDir;
 use tokio::{net::TcpListener, runtime::Runtime};
+
+use crate::storage_options::{azure_storage_options_from_env, s3_storage_options_from_env};
 
 const S3S_ACCESS_KEY: &str = "AKIAIOSFODNN7EXAMPLE";
 const S3S_SECRET_KEY: &str = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY";
@@ -210,38 +211,6 @@ fn azure_container_env() -> Option<String> {
 
 fn azure_prefix_root(default: &str) -> String {
     std::env::var("INFINO_REAL_AZURE_PREFIX").unwrap_or_else(|_| default.to_string())
-}
-
-/// Collect object-store credential options from `env_to_key` pairs that
-/// are set. Infino's providers no longer read the environment; the bench
-/// harness gathers credentials here and passes them as config.
-fn storage_options_from_env(env_to_key: &[(&str, &str)]) -> HashMap<String, String> {
-    env_to_key
-        .iter()
-        .filter_map(|(env, key)| std::env::var(env).ok().map(|v| (key.to_string(), v)))
-        .collect()
-}
-
-/// Standard S3 credential options from the AWS environment.
-/// `AWS_DEFAULT_REGION` is listed before `AWS_REGION` so the latter wins
-/// when both are set (matching AWS precedence; equal keys, last wins).
-pub fn s3_storage_options_from_env() -> HashMap<String, String> {
-    storage_options_from_env(&[
-        ("AWS_ACCESS_KEY_ID", "aws_access_key_id"),
-        ("AWS_SECRET_ACCESS_KEY", "aws_secret_access_key"),
-        ("AWS_SESSION_TOKEN", "aws_session_token"),
-        ("AWS_DEFAULT_REGION", "aws_region"),
-        ("AWS_REGION", "aws_region"),
-        ("AWS_ENDPOINT", "aws_endpoint"),
-    ])
-}
-
-/// Standard Azure credential options from the environment.
-pub fn azure_storage_options_from_env() -> HashMap<String, String> {
-    storage_options_from_env(&[
-        ("AZURE_STORAGE_ACCOUNT_NAME", "azure_storage_account_name"),
-        ("AZURE_STORAGE_ACCOUNT_KEY", "azure_storage_account_key"),
-    ])
 }
 
 /// Whether to retain the run's unique prefix instead of deleting it.
